Refactor login to work properly.
