+++ src/css/_breakpoints.scss
@@ -20,7 +20,7 @@
       font-size: 1.5rem;
     }
 
-    .trip-preview, .destination-preview {
+    .trip-preview, .dashboard-message {
       grid-area: auto;
     }
 
@@ -46,7 +46,7 @@
       font-size: 1rem;
     }
 
-    .trip-preview, .destination-preview {
+    .trip-preview, .dashboard-message {
       grid-area: 4;
     }
